Create a class for the RSC representation
Also create classes for dsCMatrix, CHMfactor, etc in the hope that
they can be converted to CHM_SP, etc. and generated from same.  Not
done with that yet.

Also, the current commit gets the wrong answers because I have
switched to having rv just provide the row indices for the Z part of
the matrix as the other rows are redundant.  However some of the code
in the update_A method must be updated.