Add argument to simulate() to allow user-specified distn cluster random effects.
* modify simulate.merMod() and .simulateFun() to take user-specified cluster random effects generator.
* update documentation
* add unit test for new feature
* add vignette simulation.Rmd describing the old code and discussing the design of this modification.
* update version and add self as contributor

The unit test is an addition tests/simulate.R, and so is not a testthat test.
I think this mean devtools::test will not trigger it, though devtools::check does.
Also, the test will *not* run unless LME4_TEST_LEVEL >= 2, and so ordinarily
it will not run--same as all the tests in simulate.R.

In running this I generated a new tests/testthat/test-summary_testleve1_1.rda.
Although the file is under version control, I have not committed the changes.  It also
generated a ..._testlevel_2.rda file, which I left not under version control.

When I raise LME4_TEST_LEVEL to 2 it generates a lot of test failures in
code I didn't touch.  However, the new test does pass.

The "vignette" I added isn't really intended to be user facing.